Bug fix: The test to determine if VBoxManage is installed failed if VBoxManage was not in fact installed!
Bug fix: If VBoxManage was not installed the Preferences dialog crashed. 

GTK.AboutDialog now uses a Stack rather than Notebook to hold the underlying widgets.  Rather than try to (constantly) reverse engineer the GTK.AboutDialog to retrofit the changelog tab, the default GTK.AboutDialog will be used and the changelog will be shown in a final tab of the Preferences dialog.