Bug fix: The Icon Text was incorrectly parsing nested brackets.
Bug fix: The declination attribute was incorrectly including the '-' when the value was South.

GTK.AboutDialog now uses a Stack rather than Notebook to hold the underlying widgets.  Rather than try to (constantly) reverse engineer the GTK.AboutDialog to retrofit the changelog tab, the default GTK.AboutDialog is used with a hyperlink to the changelog inserted.

Removed all display formatting from backend and put into frontend.