Added URL to list of references.
Removed TODO: Need to kick of update() from main() in a new thread (via GLib.idle_add) as the update() could take a while and block the main loop and therefore the GUI blocks and/or becomes unresponsive.

Combined the onAbout() and onAboutInternal() into one function; makes no differnce in terms of timing or GUI blocking with two functions.